% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{death_by_time}
\alias{death_by_time}
\title{COVID-19 weekly number of death in Norway}
\format{
\describe{
\item{granularity_time}{Time granularity: isoweek}
\item{granularity_geo}{Geographical granularity: nation}
\item{location_code}{Location code: norge}
\item{border}{Municipality border: 2020}
\item{age}{Age category: total}
\item{sex}{Sex category: total}
\item{year}{Year of data registration}
\item{week}{ISO week of data registration}
\item{yrwk}{Year week of data registration}
\item{season}{Influenza season}
\item{x}{Season week}
\item{date}{Date of data registration}
\item{n}{Number of cases}
\item{date_of_publishing}{Date of data release: 2022-11-11}
}
}
\usage{
death_by_time
}
\description{
This dataset contains weekly number of death in Norway,
from 2020 week 8 (2020-02-23) to 2022 week 45 (2022-11-13)
}
\keyword{datasets}
