% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{demographics}
\alias{demographics}
\title{COVID-19 total number of death in Norway by age and sex}
\format{
\describe{
\item{granularity_time}{Time granularity: total}
\item{granularity_geo}{Geographical granularity: nation}
\item{location_code}{Location code: norge}
\item{border}{Municipality border. All 2020}
\item{age}{Age category: 0-39, 40-49, 50-59, 60-69, 70-79, 80-89, 90+}
\item{sex}{Sex category: male, female}
\item{year}{1900, i.e. of all time}
\item{week}{1, i.e. of all time}
\item{yrwk}{1900-01, i.e. of all time}
\item{season}{Influenza season}
\item{x}{Season week}
\item{date}{1900-01-01, i.e. of all time}
\item{n}{Number of cases}
\item{date_of_publishing}{Date of data release}
\item{tag_outcome}{Outcome tag. Death}
}
}
\usage{
demographics
}
\description{
This dataset contains total number of death in Norway caused by COVID-19
}
\keyword{datasets}
