% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_loaddata.R
\name{dummy_fn}
\alias{dummy_fn}
\title{Dummy function placeholder}
\usage{
dummy_fn(x, y)
}
\arguments{
\item{x}{a number}

\item{y}{another number}
}
\value{
the sum
}
\description{
Dummy function placeholder
}
\examples{
x <- 3
y <- 4

z <- dummy_fn(x, y)
}
