% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{hospital_by_time}
\alias{hospital_by_time}
\title{COVID-19 daily number of hospitalisation and ICU admissions in Norway}
\format{
\describe{
\item{granularity_time}{Time granularity: day}
\item{granularity_geo}{Geographical granularity: nation}
\item{location_code}{Location code: norge}
\item{border}{Municipality border: 2020}
\item{age}{Age category: total}
\item{sex}{Sex category: total}
\item{year}{Year of data registration}
\item{week}{ISO week of data registration}
\item{yrwk}{Year week of data registration}
\item{season}{Influenza season}
\item{x}{Season week}
\item{date}{Date of data registration}
\item{n_icu}{Number of ICU admissions}
\item{n_hospital_main_cause}{Number of hospitalisation where COVID-19 is the main cause}
\item{date_of_publishing}{Date of data release}
}
}
\usage{
hospital_by_time
}
\description{
This dataset contains daily number of COVID-19 hospitalisation and ICU admissions in Norway,
from 2020-02-21 to 2022-11-10
}
\keyword{datasets}
