% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{lab_by_time}
\alias{lab_by_time}
\title{COVID-19 daily number of tested in Norway}
\format{
\describe{
\item{granularity_time}{Time granularity: day}
\item{granularity_geo}{Geographical granularity: nation}
\item{location_code}{Location code: norge}
\item{border}{Municipality border: 2020}
\item{age}{Age category: total}
\item{sex}{Sex category: total}
\item{year}{Year of data registration}
\item{week}{ISO week of data registration}
\item{yrwk}{Year week of data registration}
\item{season}{Influenza season}
\item{x}{Season week}
\item{date}{Date of data registration}
\item{n_neg}{Number of negative COVID tests}
\item{n_pos}{Number of positive COVID tests}
\item{pr100_pos}{Percentage of positive tests of all tests}
\item{date_of_publishing}{Date of data release}
}
}
\usage{
lab_by_time
}
\description{
This dataset contains daily number of tested (positive, negative) in Norway,
from 2020-04-01 to 2022-11-10
}
\keyword{datasets}
