% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{msis_by_location}
\alias{msis_by_location}
\title{MSIS confirmed COVID-19 cases of all time in Norway}
\format{
\describe{
\item{location_code}{Location code in Norway}
\item{date}{1900-01-01, i.e. of all time}
\item{n}{Number of cases}
\item{granularity_time}{Time granularity: total}
\item{granularity_geo}{Geographical granularity: nation, county, municip}
\item{border}{Municipality border: 2020}
\item{age}{Age category: total}
\item{sex}{Sex category: total}
\item{year}{1900, i.e. of all time}
\item{week}{1, i.e. of all time}
\item{yrwk}{1900-01, i.e. of all time}
\item{season}{Influenza season}
\item{x}{Season week}
\item{date}{1900-01-01, i.e. of all time}
\item{pop}{Population}
\item{pr100000}{Number of cases per 100 000 population}
\item{location_name}{Location name, matches location code}
\item{date_of_publishing}{Date of data release}
}
}
\usage{
msis_by_location
}
\description{
This dataset contains MSIS confirmed COVID-19 cases of all time in national, county and municipality levels.
}
\keyword{datasets}
