% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{msis_by_time_location}
\alias{msis_by_time_location}
\title{Daily MSIS confirmed COVID-19 cases in Norway}
\format{
\describe{
\item{location_code}{Location code in Norway}
\item{date}{Date of data registration}
\item{n}{Number of cases}
\item{granularity_time}{Time granularity: total}
\item{granularity_geo}{Geographical granularity: nation, county, municip}
\item{border}{Municipality border: 2020}
\item{age}{Age category: total}
\item{sex}{Sex category: total}
\item{year}{Year of data registration}
\item{week}{ISO week of data registration}
\item{yrwk}{Year week of data registration}
\item{season}{Influenza season}
\item{x}{Season week}
\item{pop}{Population}
\item{pr100000}{Number of cases per 100 000 population}
\item{location_name}{Location name, matches location code}
\item{date_of_publishing}{Date of data release}
}
}
\usage{
msis_by_time_location
}
\description{
This dataset contains daily MSIS confirmed COVID-19 cases in national, county and municipality levels,
from 2020-02-21 to 2022-11-10
}
\keyword{datasets}
