% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{sysvak_by_location_sex_age}
\alias{sysvak_by_location_sex_age}
\title{Vaccination in Norway of all time by sex and age}
\format{
\describe{
\item{granularity_time}{Time granularity: total}
\item{granularity_geo}{Geographical granularity: nation, county, municip}
\item{location_code}{Location code in Norway}
\item{border}{Municipality border: 2020}
\item{age}{Age category: 12-15, 16-17, 18-24, 25-39, 40-44, 45-54, 55-64, 65-74, 75-79, 80-84, above 85}
\item{sex}{Sex category: male, female}
\item{year}{1900, i.e. of all time}
\item{week}{1, i.e. of all time}
\item{yrwk}{1900-01, i.e. of all time}
\item{season}{Influenza season}
\item{x}{Season week}
\item{date}{1900-01-01, i.e. of all time}
\item{n_dose_1}{Number of dose 1}
\item{n_dose_2}{Number of dose 2}
\item{n_dose_3}{Number of dose 3}
\item{n_dose_4}{Number of dose 4}
\item{pr100_dose_1}{Percentage of dose 1}
\item{pr100_dose_2}{Percentage of dose 2}
\item{pr100_dose_3}{Percentage of dose 3}
\item{pr100_dose_4}{Percentage of dose 4}
\item{pop}{Population}
\item{location_name}{Location name, matches location code}
\item{date_of_publishing}{Date of data release}
}
}
\usage{
sysvak_by_location_sex_age
}
\description{
This dataset contains total vaccination of 1st, 2nd, 3rd and 4th dose in national, county and municipality levels, categorised by age groups and sex.
}
\keyword{datasets}
