% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{sysvak_by_time_location}
\alias{sysvak_by_time_location}
\title{Daily vaccination in Norway}
\format{
\describe{
\item{granularity_time}{Time granularity: day}
\item{granularity_geo}{Geographical granularity: nation, county, municip}
\item{location_code}{Location code in Norway}
\item{border}{Municipality border: 2020}
\item{age}{Age category: total}
\item{sex}{Sex category: total}
\item{year}{Year of data registration}
\item{week}{ISO week of data registration}
\item{yrwk}{Year week of data registration}
\item{season}{Influenza season}
\item{x}{Season week}
\item{n_dose_1}{Number of dose 1}
\item{n_dose_2}{Number of dose 2}
\item{n_dose_3}{Number of dose 3}
\item{n_dose_4}{Number of dose 4}
\item{cum_n_dose_1}{Cumulative number of dose 1}
\item{cum_n_dose_2}{Cumulative number of dose 2}
\item{cum_n_dose_3}{Cumulative number of dose 3}
\item{cum_n_dose_4}{Cumulative number of dose 4}
\item{cum_pr100_dose_1}{Cumulative percentage of dose 1}
\item{cum_pr100_dose_2}{Cumulative percentage of dose 2}
\item{cum_pr100_dose_3}{Cumulative percentage of dose 3}
\item{cum_pr100_dose_4}{Cumulative percentage of dose 4}
\item{pop}{Population}
\item{location_name}{Location name, matches location code}
\item{date_of_publishing}{Date of data release}
}
}
\usage{
sysvak_by_time_location
}
\description{
This dataset contains daily and cumulative vaccination of 1st, 2nd, 3rd and 4th dose in national, county and municipality levels,
from 2020-12-02 to 2022-11-10.
}
\keyword{datasets}
